A lot of stuff

+++ src/main/java/com/miausoft/miaups/persistence/ParcelDimensionsRepository.java
@@ -4,5 +4,5 @@ package com.miausoft.miaups.persistence;
 import com.miausoft.miaups.persistence.entities.ParcelDimensions;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ParcelDimensionsRepository extends JpaRepository<ParcelDimensions, Integer> {
+public interface ParcelDimensionsRepository extends JpaRepository<ParcelDimensions, Long> {
 }
