+++ src/main/java/com/miausoft/miaups/persistence/ParcelMachinesRepository.java
@@ -3,5 +3,5 @@ package com.miausoft.miaups.persistence;
 import com.miausoft.miaups.persistence.entities.ParcelMachine;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ParcelMachinesRepository extends JpaRepository<ParcelMachine,Integer> {
+public interface ParcelMachinesRepository extends JpaRepository<ParcelMachine,Long> {
 }
